Add Component Rank-Size Distribution and refactoring Move partitioning plots into own submodule Including tests
